moved shell task into common namespace
