+++ arcana/core/tasks/__init__.py
@@ -1,2 +1,2 @@
-from .shell import shell_cmd
+from ...common.tasks.shell import shell_cmd
 from .utils import identity, identity_converter
